requirements.txt
# Vulnerable versions for testing pip-audit
urllib3==1.26.16       # <1.26.17 affected
# Pillow==10.1.0         # <=10.1.0 affected
PyYAML==4.2b4            # <5.1 affected
Django==4.2.24         # <4.2.25 affected

